ReadToMe-backend/backend/requirements.txt
argcomplete==1.10.3
asgiref==3.7.2
beautifulsoup4==4.8.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
compressed-rtf==1.0.6
Django==4.2.5
django-cors-headers==4.2.0
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
gTTS==2.3.2
idna==3.4
IMAPClient==2.1.0
lxml==4.9.3
olefile==0.46
packaging==23.2
pdfminer.six==20191110
Pillow==9.5.0
pycryptodome==3.19.0
python-pptx==0.6.22
requests==2.31.0
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
sqlparse==0.4.4
textract==1.6.5
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.6
xlrd==1.2.0
XlsxWriter==3.1.6
